+++ .gitignore
@@ -9,3 +9,4 @@ jobsim.dSYM/
 test
 benchmark
 Benchmark/
+swapfile
