+++ asst2/my_malloc.c
@@ -1,9 +1,12 @@
+#include <string.h>
+
 #include "my_malloc.h"
 #include "../asst1/my_pthread_t.h"
 #include "./virtual_memory.h"
 
 /* Globals */
 extern int is_initialized;
+extern FILE *swapfile;
 
 /* Misc */
 
@@ -12,15 +15,6 @@ int my_ceil(double num){
         return (int)num;
     }
     return (int)num + 1;
-}
-
-/* mprotect handler */
-
-void handler(int sig, siginfo_t *si, void *unused) {
-    intptr_t offset = (intptr_t)si->si_addr - (intptr_t)((void*) allmem);
-    int page_num = offset / PAGE_SIZE;
-    printf("\nERROR TRYING TO ACCESS PAGE #%d as thread %d\n", page_num, get_curr_tcb_id());
-    exit(1);
 }
 
 void load_pages(int id) {
@@ -253,182 +247,6 @@ void set_PTE_location(int id, int idx, int location) {
         }
 
         cur = cur->next;
-    }
-}
-
-/* Memory Initialization */
-
-void *create_mdata(){
-    printf("Creating page meta data \n");
-    int pages_needed = MDATA_NUM_PAGES;
-
-    printf("NUM PAGES FOR MDATA %d\n", pages_needed);
-
-    int i;
-    int mdata_start_idx = NUM_PAGES-pages_needed;
-    // Start mdata at totalnumpages-(mdata)
-
-    printf("num pages: %d          mdata size: %lu\n", MDATA_NUM_PAGES, MDATA_SIZE);
-    // make an Entry for mdata
-    Entry *mdata_entry = (Entry*) (allmem + ((NUM_PAGES - pages_needed) * PAGE_SIZE));//((char*)SYSINFO + PAGE_SIZE);
-
-    Page * root = (Page*) (mdata_entry + 1);
-    Page * curr_page = root;
-    Page * prev_page = NULL;
-    
-    // Populate the page metadata (in OS land) with all,empty pages
-    for(i = 0; i < NUM_PAGES; i++){
-        // mdata
-        if (i >= mdata_start_idx) {
-            curr_page->id = 0;
-            curr_page->is_free = 0;
-            curr_page->mem_free = PAGE_SIZE;
-            curr_page->next = (Page *) ( (char *)curr_page + PAGE_STRUCT_SIZE);
-            curr_page->prev = prev_page;
-            curr_page->front = GET_PAGE_ADDRESS(i);
-            curr_page->idx = i;
-            curr_page->cur_idx = i;
-            curr_page->parent = NUM_PAGES - MDATA_NUM_PAGES;
-        } else if (i == mdata_start_idx-1) { // new page where other sys stuff goes
-            curr_page->id = 0;
-            curr_page->is_free = 0;
-            curr_page->mem_free = PAGE_SIZE;
-            curr_page->next = (Page *) ( (char *)curr_page + PAGE_STRUCT_SIZE);
-            curr_page->prev = prev_page;
-            curr_page->front = GET_PAGE_ADDRESS(i);
-            curr_page->idx = -1;
-            curr_page->cur_idx = i;
-            curr_page->parent = i;
-        } else {
-            curr_page->id = -1;
-            curr_page->is_free = 1;
-            curr_page->mem_free = PAGE_SIZE;
-            curr_page->next = (Page *) ( (char *)curr_page + PAGE_STRUCT_SIZE);
-            curr_page->prev = prev_page;
-            curr_page->front = GET_PAGE_ADDRESS(i);
-            curr_page->idx = -1;
-            curr_page->cur_idx = i;
-            curr_page->parent = -1;
-        }
-
-        if (i <= mdata_start_idx) {
-            init_front(curr_page);
-            if (i == mdata_start_idx || i == (mdata_start_idx - 1)) {
-                curr_page->front->is_free = 0;
-            }
-        }
-
-        prev_page = curr_page;
-        curr_page = curr_page->next;
-    }
-
-    // init mdata for swap file
-    for (; i < NUM_PAGES + NUM_SWAPFILE_PAGES; i++) {
-        curr_page->id = -1;
-        curr_page->is_free = 1;
-        curr_page->mem_free = PAGE_SIZE;
-        curr_page->next = (Page *) ( (char *)curr_page + PAGE_STRUCT_SIZE);
-        curr_page->prev = prev_page;
-        curr_page->front = GET_PAGE_ADDRESS(i);
-        curr_page->idx = -1;
-        curr_page->cur_idx = i;
-        curr_page->parent = -1;
-    }
-
-    // init swapfile pages
-    for (i = 0; i < NUM_SWAPFILE_PAGES; i++) {
-        int swap_offset = i * PAGE_SIZE;
-
-        // from temp to swap
-        fseek(swapfile, swap_offset, SEEK_SET);
-        int read = fwrite(GET_PAGE_ADDRESS(0), PAGE_SIZE, 1, swapfile);
-    }
-
-    *mdata_entry = (Entry) {
-        .size = MDATA_SIZE,
-            .next = NULL,
-            .is_free = 0 };
-    curr_page->next = NULL;
-
-    return (void *)((char *)SYSINFO - sizeof(Entry));
-}
-
-void *create_pagetable(void * end_of_mdata){
-    printf("Creating page table and SysInfo \n");
-    int prev_remaining_space = PAGE_SIZE - sizeof(Entry);
-
-    // SysInfo
-    Entry *sys_info_entry = ((Entry*)end_of_mdata);
-    *sys_info_entry = (Entry) {
-        .size = sizeof(SysInfo),
-        .is_free = 0,
-        .next = (Entry*) (((char*)(sys_info_entry + 1)) + sizeof(SysInfo)) };
-    SysInfo *info = (SysInfo*) (sys_info_entry + 1);
-    *info = (SysInfo) {
-        .pagetable = (PTE**) (sys_info_entry->next + 1),
-        .mdata = (Page*) ((char *)SYSINFO + PAGE_SIZE) };
-
-    // page_table outer
-    Entry *page_table_outer_entry = sys_info_entry->next;
-    *page_table_outer_entry = (Entry) {
-        .size = sizeof(PTE*),
-        .is_free = 0,
-        .next = (Entry*) (((char*)(page_table_outer_entry + 1)) + sizeof(PTE*)) };
-    PTE **page_table_outer = (PTE**) (page_table_outer_entry + 1);
-    *page_table_outer = (PTE*) (((char*)(page_table_outer + 1)) + sizeof(Entry));
-
-    // page_table inner
-    int num_sys_pages = my_ceil((double)MDATA_SIZE / (double)PAGE_SIZE) + 1;
-    Entry *page_table_inner_entry = page_table_outer_entry->next;
-    *page_table_inner_entry = (Entry) {
-        .size = sizeof(PTE) * num_sys_pages,
-        .is_free = 0,
-        .next = (Entry*) (((char*)(page_table_inner_entry + 1)) + (sizeof(PTE) * num_sys_pages)) };
-    int i = 0;
-    for (; i < num_sys_pages; i++) {
-        page_table_outer[0][i] = (PTE) {
-            .page_index = i,
-            .page_loc = i,
-            .next = (i < num_sys_pages - 1) ? (PTE*) (&page_table_outer[0][0] + (i+1)) : NULL };
-    }
-
-    // remaining space entry
-    Entry *remaining_entry = page_table_inner_entry->next;
-    int size_of_everything_we_made = sizeof(SysInfo) + sizeof(Entry) + sizeof(PTE*) + sizeof(Entry) + sizeof(PTE) + sizeof(Entry);
-    *remaining_entry = (Entry) {
-        .size = prev_remaining_space - size_of_everything_we_made,
-        .is_free = 1,
-        .next = NULL };
-
-    return (void*)(remaining_entry + 1);
-}
-
-void make_swapfile() {
-    int num_bytes = SWAPFILE_SIZE;
-    swapfile = fopen("swapfile", "w+");
-    fseek(swapfile, num_bytes, SEEK_SET);
-    fputc('\0', swapfile);
-}
-
-void mem_init(){
-    printf("MEM INIT\n");
-
-    posix_memalign((void**)&allmem, PAGE_SIZE, ARRAY_SIZE);
-
-    make_swapfile();
-
-    void * end_of_mdata = create_mdata();
-    create_pagetable(end_of_mdata);
-
-    //register segfault handler
-    struct sigaction sa;
-    sa.sa_flags = SA_SIGINFO;
-    sigemptyset(&sa.sa_mask);
-    sa.sa_sigaction = handler;
-
-    if (sigaction(SIGSEGV, &sa, NULL) == -1) {
-        printf("ERROR CANT SETUP SEGFAULT HANDLER\n");
-        exit(1);
     }
 }
